requirements.txt
evdev>=1.6.1
pyyaml>=6.0.1
python-xlib>=0.33
xdotool>=0.1.0 